use standard int types in header

+++ source/arm9/wifi_arm9.c
@@ -1059,7 +1059,7 @@ void Wifi_Timer(int num_ms)
 
 #endif
 
-unsigned long Wifi_Init(int initflags)
+u32 Wifi_Init(int initflags)
 {
     if (Wifi_Data_Struct == NULL)
     {
